The simple subset of changes recommended by Ioan Sucan.
