+++ src/ompl/base/objectives/PathLengthOptimizationObjective.h
@@ -32,8 +32,7 @@
 *  POSSIBILITY OF SUCH DAMAGE.
 *********************************************************************/
 
-/* Author: Luis G. Torres */
-/* Edited by: Jonathan Gammell (allocInformedStateSampler) */
+/* Author: Luis G. Torres, Jonathan Gammell (allocInformedStateSampler) */
 
 #ifndef OMPL_BASE_OBJECTIVES_PATH_LENGTH_OPTIMIZATION_OBJECTIVE_
 #define OMPL_BASE_OBJECTIVES_PATH_LENGTH_OPTIMIZATION_OBJECTIVE_
@@ -65,7 +64,7 @@ namespace ompl
             virtual Cost motionCostHeuristic(const State *s1, const State *s2) const;
 
             /** \brief Allocate a state sampler for the path-length objective (i.e., direct ellipsoidal sampling). */
-            virtual InformedStateSamplerPtr allocInformedStateSampler(const StateSpace* space, const ProblemDefinitionPtr probDefn, const Cost* bestCost) const;
+            virtual InformedStateSamplerPtr allocInformedStateSampler(const StateSpace* space, const ProblemDefinitionPtr probDefn, const GetCurrentCost& costFunc) const;
         };
     }
 }
